pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comcan-ctx"
version = "0.1.0"
description = "Context management for AI coding agents — state awareness + expertise memory."
authors = [
    {name = "Shriniwas Ramesh Suram"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
keywords = ["ai", "context", "cursor", "coding-agent", "llm", "git"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tiktoken>=0.5.0",
    "pathspec>=0.11.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "bandit>=1.7",
]

[project.scripts]
comcan = "comcan.cli:app"

[tool.setuptools.packages.find]
include = ["comcan*"]

[tool.bandit]
exclude_dirs = ["tests"]
